refactor names of libs and update readme
