+++ README.md
@@ -1,26 +1,6 @@
 # stem skeleton
 
-Cell with predefined dna having the following abilities:
-
-* dual architecture - api & client site
-* expressjs server at port 1337 configured with
-  * body-parser
-  * cookie-parser
-  * views at /context/pages
-  * uploads handling
-  * connect-mongo sessions
-  * api & site responders
-  * js/css cache busting based on project's version
-* autoload and mount api action handlers from /context/routes/api
-* autoload and mount site action handlers from /context/routes/site
-* autoload and mount static pages from /context/pages
-* in development 
-  * watch all javascripts and bundle them from /context/client to /public/js
-  * watch all stylesheets and bundle them from /context/styles to /public/css
-  * serve static files from /public and /build folders to /public url
-  * continuous delivery
-    * development, test and staging modes
-    * full remote process management (install, build, start, restart, upgrade, uninstall)
+Cell with predefined dna to be developed as 'backend', 'frontend' or both
 
 ## usage
 
@@ -28,32 +8,46 @@ Cell with predefined dna having the following abilities:
 - `cd ./mypetproject`
 - `rm -rf ./.git`
 - `npm install`
-- `node index.js`
+- modify `dna` folder and `package.json`
 
-### optional perks
+### stack addons
 
-#### use jade templates and backbone/jquery on the frontend
+#### server
 
-    $ node ./node_modules/.bin/angel stack add upgrades/jade-backbone
+##### emails support
 
-#### use ejs serverside templates and angularjs
+    $ node ./node_modules/.bin/angel stack add upgrades/server/emails-support
 
-    $ node ./node_modules/.bin/angel stack add upgrades/ejs-angular
+##### mongoose models
 
-#### use reactJS with flux on the frontend
+    $ node ./node_modules/.bin/angel stack add upgrades/server/mongoose
 
-    $ node ./node_modules/.bin/angel stack add upgrades/react-flux
+##### mongodb stored cookie based sessions
 
-#### use jade email templates and support email delivery via plasma
+    $ node ./node_modules/.bin/angel stack add upgrades/server/mongo-sessions
 
-    $ node ./node_modules/.bin/angel stack add upgrades/emails-support
+#### client
 
-### cleanup before initial commit
+##### server rendered pages
 
-* Delete the following files manually once not needed:
+###### ejs templates
 
-  * `context/pages/**/*.jade`
-  * `upgrades` folder
+    $ node ./node_modules/.bin/angel stack add upgrades/client/ejs-pages
 
-* Modify `dna` folder contents accordingly to your needs.
-* Modify `package.json` contents accordingly to your needs.
+###### jade templates
+
+    $ node ./node_modules/.bin/angel stack add upgrades/client/jade-pages
+
+##### asset pipeline
+
+###### browserify
+
+    $ node ./node_modules/.bin/angel stack add upgrades/client/assetpipeline-browserify
+
+###### webpack
+
+    $ node ./node_modules/.bin/angel stack add upgrades/client/assetpipeline-webpack
+
+###### less
+
+    $ node ./node_modules/.bin/angel stack add upgrades/client/assetpipeline-less
