+++ lib/gulp-pipeline.js
