+++ lib/parallel-exec.js
