+++ upgrades/client/assetpipeline-less/scripts/watchcss.js
@@ -1,6 +1,7 @@
 module.exports = function (angel) {
   angel.on('watchcss', function (angel) {
     var loadDNA = require('../lib/dna')
+    var runPipeline = require('../lib/gulp-pipeline')
     var less = require('gulp-less')
     var lessWatcher = require('gulp-less-watcher')
 
@@ -13,7 +14,7 @@ module.exports = function (angel) {
 
     loadDNA(function (err, dna) {
       var options = dna.client.assetpipeline
-      require('../lib/angelabilities-gulp-stream')({
+      runPipeline({
         name: 'watchcss',
         src: options.src + '/**/*.bundle.css',
         pipeline: [
