+++ upgrades/client/assetpipeline-webpack/scripts/watchjs.js
@@ -1,12 +1,13 @@
 module.exports = function(angel) {
   angel.on('watchjs', function () {
     var loadDNA = require('../lib/dna')
+    var runPipeline = require('../lib/gulp-pipeline')
     var webpack = require('webpack-stream')
     var sourcemaps = require('gulp-sourcemaps')
 
     loadDNA(function (err, dna) {
       var options = dna.client.assetpipeline
-      require('../lib/angelabilities-gulp-stream')({
+      runPipeline({
         name: 'watchjs',
         src: dna.client.assetpipeline.src + '/**/*.bundle.js',
         pipeline: [
